Update topk UTs
